pyproject.toml
[build-system]
requires = ["setuptools>=42.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "autonetlab"
version = "0.1.0"
description = "Automation toolkit for setting up and managing Cisco network lab environments"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Matt Facca", email = "your-email@example.com"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Network Engineers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: System :: Networking",
]

dependencies = [
    "netmiko>=4.0.0",
    "paramiko>=2.11.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "sphinx>=5.0.0",
    "sphinx-rtd-theme>=1.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[project.urls]
"Homepage" = "https://github.com/yourusername/autonetlab"
"Bug Tracker" = "https://github.com/yourusername/autonetlab/issues"

[tool.setuptools]
packages = ["autonetlab"]

[tool.black]
line-length = 88
target-version = ["py38"]
include = '\.pyi?$'

[tool.pytest.ini_options]
testpaths = ["tests"]

